Move pkiutil package into util package
No functional changes
