+++ cmd/aggregator/app/BUILD.bazel
@@ -18,9 +18,9 @@ go_library(
         "//pkg/aggregator/server/options:go_default_library",
         "//pkg/apis/aggregator/v1:go_default_library",
         "//pkg/options:go_default_library",
-        "//pkg/pkiutil:go_default_library",
         "//pkg/util:go_default_library",
         "//pkg/util/crash:go_default_library",
+        "//pkg/util/pkiutil:go_default_library",
         "//vendor/github.com/atlassian/ctrl:go_default_library",
         "//vendor/github.com/atlassian/ctrl/app:go_default_library",
         "//vendor/github.com/ghodss/yaml:go_default_library",
