+++ cmd/aggregator/app/options.go
@@ -4,7 +4,7 @@ import (
 	"io/ioutil"
 
 	"github.com/atlassian/voyager/pkg/options"
-	"github.com/atlassian/voyager/pkg/pkiutil"
+	"github.com/atlassian/voyager/pkg/util/pkiutil"
 	"github.com/ghodss/yaml"
 	"github.com/pkg/errors"
 	utilerrors "k8s.io/apimachinery/pkg/util/errors"
