+++ cmd/ops/app/BUILD.bazel
@@ -19,10 +19,10 @@ go_library(
         "//pkg/ops/informer:go_default_library",
         "//pkg/ops/server:go_default_library",
         "//pkg/ops/server/options:go_default_library",
-        "//pkg/pkiutil:go_default_library",
         "//pkg/util:go_default_library",
         "//pkg/util/crash:go_default_library",
         "//pkg/util/informers:go_default_library",
+        "//pkg/util/pkiutil:go_default_library",
         "//vendor/github.com/atlassian/ctrl:go_default_library",
         "//vendor/github.com/atlassian/ctrl/app:go_default_library",
         "//vendor/github.com/ghodss/yaml:go_default_library",
