+++ cmd/reporter/app/BUILD.bazel
@@ -27,13 +27,13 @@ go_library(
         "//pkg/options:go_default_library",
         "//pkg/orchestration/client:go_default_library",
         "//pkg/orchestration/informer:go_default_library",
-        "//pkg/pkiutil:go_default_library",
         "//pkg/reporter:go_default_library",
         "//pkg/reporter/server:go_default_library",
         "//pkg/reporter/server/options:go_default_library",
         "//pkg/util:go_default_library",
         "//pkg/util/crash:go_default_library",
         "//pkg/util/informers:go_default_library",
+        "//pkg/util/pkiutil:go_default_library",
         "//vendor/github.com/atlassian/ctrl:go_default_library",
         "//vendor/github.com/atlassian/ctrl/app:go_default_library",
         "//vendor/github.com/atlassian/smith/pkg/apis/smith/v1:go_default_library",
