+++ cmd/svccatadmission/app/BUILD.bazel
@@ -12,10 +12,10 @@ go_library(
         "//cmd:go_default_library",
         "//pkg/execution/svccatadmission:go_default_library",
         "//pkg/options:go_default_library",
-        "//pkg/pkiutil:go_default_library",
         "//pkg/util:go_default_library",
         "//pkg/util/crash:go_default_library",
         "//pkg/util/logz:go_default_library",
+        "//pkg/util/pkiutil:go_default_library",
         "//vendor/github.com/atlassian/ctrl/options:go_default_library",
         "//vendor/github.com/ghodss/yaml:go_default_library",
         "//vendor/github.com/pkg/errors:go_default_library",
