+++ cmd/synchronization/app/BUILD.bazel
@@ -15,13 +15,13 @@ go_library(
         "//pkg/k8s:go_default_library",
         "//pkg/k8s/updater:go_default_library",
         "//pkg/options:go_default_library",
-        "//pkg/pkiutil:go_default_library",
         "//pkg/releases:go_default_library",
         "//pkg/releases/deployinator/client:go_default_library",
         "//pkg/servicecentral:go_default_library",
         "//pkg/synchronization:go_default_library",
         "//pkg/util:go_default_library",
         "//pkg/util/crash:go_default_library",
+        "//pkg/util/pkiutil:go_default_library",
         "//pkg/util/prometheus:go_default_library",
         "//vendor/github.com/atlassian/ctrl:go_default_library",
         "//vendor/github.com/atlassian/ctrl/app:go_default_library",
