+++ cmd/synchronization/app/options.go
@@ -6,7 +6,7 @@ import (
 	"net/url"
 
 	"github.com/atlassian/voyager/pkg/options"
-	"github.com/atlassian/voyager/pkg/pkiutil"
+	"github.com/atlassian/voyager/pkg/util/pkiutil"
 	"github.com/ghodss/yaml"
 	"github.com/pkg/errors"
 	utilerrors "k8s.io/apimachinery/pkg/util/errors"
