+++ pkg/aggregator/BUILD.bazel
@@ -12,12 +12,12 @@ go_library(
         "//:go_default_library",
         "//pkg/apis/aggregator:go_default_library",
         "//pkg/apis/aggregator/v1:go_default_library",
-        "//pkg/pkiutil:go_default_library",
         "//pkg/util:go_default_library",
         "//pkg/util/apiservice:go_default_library",
         "//pkg/util/crash:go_default_library",
         "//pkg/util/httputil:go_default_library",
         "//pkg/util/logz:go_default_library",
+        "//pkg/util/pkiutil:go_default_library",
         "//vendor/github.com/SermoDigital/jose/jws:go_default_library",
         "//vendor/github.com/felixge/httpsnoop:go_default_library",
         "//vendor/github.com/go-chi/chi:go_default_library",
