+++ pkg/aggregator/api_handler.go
@@ -15,12 +15,12 @@ import (
 	"github.com/atlassian/voyager"
 	"github.com/atlassian/voyager/pkg/apis/aggregator"
 	agg_v1 "github.com/atlassian/voyager/pkg/apis/aggregator/v1"
-	"github.com/atlassian/voyager/pkg/pkiutil"
 	"github.com/atlassian/voyager/pkg/util"
 	"github.com/atlassian/voyager/pkg/util/apiservice"
 	"github.com/atlassian/voyager/pkg/util/crash"
 	"github.com/atlassian/voyager/pkg/util/httputil"
 	"github.com/atlassian/voyager/pkg/util/logz"
+	"github.com/atlassian/voyager/pkg/util/pkiutil"
 	"github.com/felixge/httpsnoop"
 	"github.com/go-chi/chi"
 	"github.com/pkg/errors"
