+++ pkg/creator/BUILD.bazel
@@ -18,13 +18,13 @@ go_library(
         "//pkg/creator/ssam:go_default_library",
         "//pkg/orchestration/wiring/ec2compute/common:go_default_library",
         "//pkg/pagerduty:go_default_library",
-        "//pkg/pkiutil:go_default_library",
         "//pkg/servicecentral:go_default_library",
         "//pkg/util:go_default_library",
         "//pkg/util/apiservice:go_default_library",
         "//pkg/util/auth:go_default_library",
         "//pkg/util/httputil:go_default_library",
         "//pkg/util/logz:go_default_library",
+        "//pkg/util/pkiutil:go_default_library",
         "//pkg/util/uuid:go_default_library",
         "//vendor/github.com/PagerDuty/go-pagerduty:go_default_library",
         "//vendor/github.com/pkg/errors:go_default_library",
