+++ pkg/creator/luigi/BUILD.bazel
@@ -10,9 +10,9 @@ go_library(
     importpath = "github.com/atlassian/voyager/pkg/creator/luigi",
     visibility = ["//visibility:public"],
     deps = [
-        "//pkg/pkiutil:go_default_library",
         "//pkg/util:go_default_library",
         "//pkg/util/httputil:go_default_library",
+        "//pkg/util/pkiutil:go_default_library",
         "//vendor/bitbucket.org/atlassianlabs/restclient:go_default_library",
         "//vendor/github.com/pkg/errors:go_default_library",
         "//vendor/go.uber.org/zap:go_default_library",
@@ -30,10 +30,10 @@ go_test(
     embed = [":go_default_library"],
     race = "on",
     deps = [
-        "//pkg/pkiutil:go_default_library",
         "//pkg/util:go_default_library",
         "//pkg/util/httptest:go_default_library",
         "//pkg/util/httputil:go_default_library",
+        "//pkg/util/pkiutil:go_default_library",
         "//vendor/github.com/pkg/errors:go_default_library",
         "//vendor/github.com/stretchr/testify/assert:go_default_library",
         "//vendor/github.com/stretchr/testify/mock:go_default_library",
