+++ pkg/creator/luigi/clientimpl.go
@@ -10,9 +10,9 @@ import (
 	"strings"
 
 	"bitbucket.org/atlassianlabs/restclient"
-	"github.com/atlassian/voyager/pkg/pkiutil"
 	"github.com/atlassian/voyager/pkg/util"
 	"github.com/atlassian/voyager/pkg/util/httputil"
+	"github.com/atlassian/voyager/pkg/util/pkiutil"
 	"github.com/pkg/errors"
 	"go.uber.org/zap"
 )
