+++ pkg/creator/server/options/BUILD.bazel
@@ -13,9 +13,9 @@ go_library(
         "//pkg/creator:go_default_library",
         "//pkg/creator/server/apiserver:go_default_library",
         "//pkg/pagerduty:go_default_library",
-        "//pkg/pkiutil:go_default_library",
         "//pkg/util:go_default_library",
         "//pkg/util/apiserver:go_default_library",
+        "//pkg/util/pkiutil:go_default_library",
         "//vendor/github.com/atlassian/ctrl:go_default_library",
         "//vendor/github.com/atlassian/ctrl/options:go_default_library",
         "//vendor/github.com/ghodss/yaml:go_default_library",
