+++ pkg/creator/ssam/BUILD.bazel
@@ -11,11 +11,11 @@ go_library(
     deps = [
         "//:go_default_library",
         "//pkg/creator/ssam/util/zappers:go_default_library",
-        "//pkg/pkiutil:go_default_library",
         "//pkg/ssam:go_default_library",
         "//pkg/util:go_default_library",
         "//pkg/util/httputil:go_default_library",
         "//pkg/util/logz:go_default_library",
+        "//pkg/util/pkiutil:go_default_library",
         "//pkg/util/validation:go_default_library",
         "//vendor/bitbucket.org/atlassianlabs/restclient:go_default_library",
         "//vendor/github.com/pkg/errors:go_default_library",
@@ -34,10 +34,10 @@ go_test(
     embed = [":go_default_library"],
     race = "on",
     deps = [
-        "//pkg/pkiutil/testutil:go_default_library",
         "//pkg/testutil:go_default_library",
         "//pkg/util/httptest:go_default_library",
         "//pkg/util/httputil:go_default_library",
+        "//pkg/util/pkiutil/testutil:go_default_library",
         "//vendor/github.com/stretchr/testify/assert:go_default_library",
         "//vendor/github.com/stretchr/testify/require:go_default_library",
     ],
