+++ pkg/creator/ssam/client.go
@@ -11,10 +11,10 @@ import (
 
 	"bitbucket.org/atlassianlabs/restclient"
 	"github.com/atlassian/voyager/pkg/creator/ssam/util/zappers"
-	"github.com/atlassian/voyager/pkg/pkiutil"
 	"github.com/atlassian/voyager/pkg/util"
 	"github.com/atlassian/voyager/pkg/util/httputil"
 	"github.com/atlassian/voyager/pkg/util/logz"
+	"github.com/atlassian/voyager/pkg/util/pkiutil"
 	"github.com/atlassian/voyager/pkg/util/validation"
 	"github.com/pkg/errors"
 	"go.uber.org/zap"
