+++ pkg/creator/ssam/client_test.go
@@ -8,10 +8,10 @@ import (
 	"net/url"
 	"testing"
 
-	pkitestutil "github.com/atlassian/voyager/pkg/pkiutil/testutil"
 	"github.com/atlassian/voyager/pkg/testutil"
 	. "github.com/atlassian/voyager/pkg/util/httptest"
 	"github.com/atlassian/voyager/pkg/util/httputil"
+	pkitestutil "github.com/atlassian/voyager/pkg/util/pkiutil/testutil"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
