+++ pkg/execution/svccatadmission/BUILD.bazel
@@ -17,10 +17,10 @@ go_library(
         "//:go_default_library",
         "//pkg/admission:go_default_library",
         "//pkg/execution/svccatadmission/rps:go_default_library",
-        "//pkg/pkiutil:go_default_library",
         "//pkg/servicecentral:go_default_library",
         "//pkg/util:go_default_library",
         "//pkg/util/auth:go_default_library",
+        "//pkg/util/pkiutil:go_default_library",
         "//pkg/util/uuid:go_default_library",
         "//vendor/github.com/go-chi/chi:go_default_library",
         "//vendor/github.com/kubernetes-incubator/service-catalog/pkg/apis/servicecatalog/v1beta1:go_default_library",
