+++ pkg/execution/svccatadmission/rps/BUILD.bazel
@@ -11,8 +11,8 @@ go_library(
     visibility = ["//visibility:public"],
     deps = [
         "//:go_default_library",
-        "//pkg/pkiutil:go_default_library",
         "//pkg/util:go_default_library",
+        "//pkg/util/pkiutil:go_default_library",
         "//vendor/bitbucket.org/atlassianlabs/restclient:go_default_library",
         "//vendor/github.com/pkg/errors:go_default_library",
         "//vendor/go.uber.org/zap:go_default_library",
