+++ pkg/execution/svccatadmission/rps/testing/BUILD.bazel
@@ -7,9 +7,9 @@ go_library(
     visibility = ["//visibility:public"],
     deps = [
         "//pkg/execution/svccatadmission/rps:go_default_library",
-        "//pkg/pkiutil:go_default_library",
         "//pkg/util:go_default_library",
         "//pkg/util/httptest:go_default_library",
+        "//pkg/util/pkiutil:go_default_library",
         "//vendor/github.com/stretchr/testify/require:go_default_library",
         "//vendor/go.uber.org/zap/zaptest:go_default_library",
     ],
