+++ pkg/ops/BUILD.bazel
@@ -18,11 +18,11 @@ go_library(
         "//pkg/apis/ops/v1:go_default_library",
         "//pkg/ops/constants:go_default_library",
         "//pkg/ops/util/zappers:go_default_library",
-        "//pkg/pkiutil:go_default_library",
         "//pkg/util:go_default_library",
         "//pkg/util/apiservice:go_default_library",
         "//pkg/util/httputil:go_default_library",
         "//pkg/util/logz:go_default_library",
+        "//pkg/util/pkiutil:go_default_library",
         "//pkg/util/prometheus:go_default_library",
         "//vendor/bitbucket.org/atlassianlabs/restclient:go_default_library",
         "//vendor/github.com/atlassian/ctrl:go_default_library",
@@ -56,9 +56,9 @@ go_test(
         "//pkg/apis/ops/v1:go_default_library",
         "//pkg/k8s:go_default_library",
         "//pkg/ops/constants:go_default_library",
-        "//pkg/pkiutil:go_default_library",
         "//pkg/util/httptest:go_default_library",
         "//pkg/util/logz:go_default_library",
+        "//pkg/util/pkiutil:go_default_library",
         "//vendor/github.com/SermoDigital/jose/jws:go_default_library",
         "//vendor/github.com/ash2k/stager:go_default_library",
         "//vendor/github.com/go-chi/chi:go_default_library",
