+++ pkg/ops/api_handler.go
@@ -13,10 +13,10 @@ import (
 	"github.com/atlassian/voyager/pkg/apis/ops"
 	"github.com/atlassian/voyager/pkg/ops/constants"
 	"github.com/atlassian/voyager/pkg/ops/util/zappers"
-	"github.com/atlassian/voyager/pkg/pkiutil"
 	"github.com/atlassian/voyager/pkg/util/apiservice"
 	"github.com/atlassian/voyager/pkg/util/httputil"
 	"github.com/atlassian/voyager/pkg/util/logz"
+	"github.com/atlassian/voyager/pkg/util/pkiutil"
 	prom_util "github.com/atlassian/voyager/pkg/util/prometheus"
 	"github.com/felixge/httpsnoop"
 	"github.com/go-chi/chi"
