+++ pkg/ops/api_handler_test.go
@@ -15,8 +15,8 @@ import (
 	"github.com/atlassian/voyager/pkg/apis/ops"
 	"github.com/atlassian/voyager/pkg/k8s"
 	c "github.com/atlassian/voyager/pkg/ops/constants"
-	"github.com/atlassian/voyager/pkg/pkiutil"
 	"github.com/atlassian/voyager/pkg/util/logz"
+	"github.com/atlassian/voyager/pkg/util/pkiutil"
 	"github.com/go-chi/chi"
 	"github.com/go-chi/chi/middleware"
 	sc_v1b1 "github.com/kubernetes-incubator/service-catalog/pkg/apis/servicecatalog/v1beta1"
