+++ pkg/ops/provider.go
@@ -13,10 +13,10 @@ import (
 	ctrllogz "github.com/atlassian/ctrl/logz"
 	ops_v1 "github.com/atlassian/voyager/pkg/apis/ops/v1"
 	"github.com/atlassian/voyager/pkg/ops/util/zappers"
-	"github.com/atlassian/voyager/pkg/pkiutil"
 	"github.com/atlassian/voyager/pkg/util"
 	"github.com/atlassian/voyager/pkg/util/apiservice"
 	"github.com/atlassian/voyager/pkg/util/logz"
+	"github.com/atlassian/voyager/pkg/util/pkiutil"
 	"github.com/pkg/errors"
 	"go.uber.org/zap"
 	"k8s.io/apimachinery/pkg/runtime"
