+++ pkg/reporter/BUILD.bazel
@@ -20,13 +20,13 @@ go_library(
         "//pkg/k8s:go_default_library",
         "//pkg/ops:go_default_library",
         "//pkg/ops/util/zappers:go_default_library",
-        "//pkg/pkiutil:go_default_library",
         "//pkg/util:go_default_library",
         "//pkg/util/apiservice:go_default_library",
         "//pkg/util/crash:go_default_library",
         "//pkg/util/httputil:go_default_library",
         "//pkg/util/layers:go_default_library",
         "//pkg/util/logz:go_default_library",
+        "//pkg/util/pkiutil:go_default_library",
         "//vendor/github.com/atlassian/ctrl/apis/condition/v1:go_default_library",
         "//vendor/github.com/atlassian/smith:go_default_library",
         "//vendor/github.com/atlassian/smith/pkg/apis/smith/v1:go_default_library",
@@ -74,8 +74,8 @@ go_test(
         "//pkg/ops:go_default_library",
         "//pkg/orchestration/client/fake:go_default_library",
         "//pkg/orchestration/informer:go_default_library",
-        "//pkg/pkiutil:go_default_library",
         "//pkg/util/logz:go_default_library",
+        "//pkg/util/pkiutil:go_default_library",
         "//vendor/github.com/SermoDigital/jose/jws:go_default_library",
         "//vendor/github.com/ash2k/stager:go_default_library",
         "//vendor/github.com/atlassian/ctrl/apis/condition/v1:go_default_library",
