+++ pkg/reporter/api_handler.go
@@ -19,12 +19,12 @@ import (
 	"github.com/atlassian/voyager/pkg/k8s"
 	"github.com/atlassian/voyager/pkg/ops"
 	opsZappers "github.com/atlassian/voyager/pkg/ops/util/zappers"
-	"github.com/atlassian/voyager/pkg/pkiutil"
 	"github.com/atlassian/voyager/pkg/util/apiservice"
 	"github.com/atlassian/voyager/pkg/util/crash"
 	"github.com/atlassian/voyager/pkg/util/httputil"
 	"github.com/atlassian/voyager/pkg/util/layers"
 	"github.com/atlassian/voyager/pkg/util/logz"
+	"github.com/atlassian/voyager/pkg/util/pkiutil"
 	"github.com/felixge/httpsnoop"
 	"github.com/go-chi/chi"
 	"github.com/pkg/errors"
