+++ pkg/reporter/namespace_report_test.go
@@ -19,8 +19,8 @@ import (
 	reporter_v1 "github.com/atlassian/voyager/pkg/apis/reporter/v1"
 	"github.com/atlassian/voyager/pkg/k8s"
 	"github.com/atlassian/voyager/pkg/ops"
-	"github.com/atlassian/voyager/pkg/pkiutil"
 	"github.com/atlassian/voyager/pkg/util/logz"
+	"github.com/atlassian/voyager/pkg/util/pkiutil"
 	sc_v1b1 "github.com/kubernetes-incubator/service-catalog/pkg/apis/servicecatalog/v1beta1"
 	"github.com/stretchr/testify/require"
 	"go.uber.org/zap/zaptest"
