+++ pkg/servicecentral/BUILD.bazel
@@ -15,10 +15,10 @@ go_library(
     deps = [
         "//:go_default_library",
         "//pkg/apis/creator/v1:go_default_library",
-        "//pkg/pkiutil:go_default_library",
         "//pkg/util:go_default_library",
         "//pkg/util/auth:go_default_library",
         "//pkg/util/httputil:go_default_library",
+        "//pkg/util/pkiutil:go_default_library",
         "//vendor/bitbucket.org/atlassianlabs/restclient:go_default_library",
         "//vendor/github.com/pkg/errors:go_default_library",
         "//vendor/go.uber.org/zap:go_default_library",
@@ -42,13 +42,13 @@ go_test(
     deps = [
         "//:go_default_library",
         "//pkg/apis/creator/v1:go_default_library",
-        "//pkg/pkiutil:go_default_library",
-        "//pkg/pkiutil/testutil:go_default_library",
         "//pkg/testutil:go_default_library",
         "//pkg/util:go_default_library",
         "//pkg/util/auth:go_default_library",
         "//pkg/util/httptest:go_default_library",
         "//pkg/util/httputil:go_default_library",
+        "//pkg/util/pkiutil:go_default_library",
+        "//pkg/util/pkiutil/testutil:go_default_library",
         "//vendor/github.com/pkg/errors:go_default_library",
         "//vendor/github.com/stretchr/testify/assert:go_default_library",
         "//vendor/github.com/stretchr/testify/mock:go_default_library",
