+++ pkg/servicecentral/client.go
@@ -10,10 +10,10 @@ import (
 	"time"
 
 	"bitbucket.org/atlassianlabs/restclient"
-	"github.com/atlassian/voyager/pkg/pkiutil"
 	"github.com/atlassian/voyager/pkg/util"
 	"github.com/atlassian/voyager/pkg/util/auth"
 	"github.com/atlassian/voyager/pkg/util/httputil"
+	"github.com/atlassian/voyager/pkg/util/pkiutil"
 	"github.com/pkg/errors"
 	"go.uber.org/zap"
 )
